Remove the reference to openshift/api
The referenced tag has been removed, which breaks dependabot (#791).

This is another attempt to fix it, by removing an explicit reference
(which was added when updating Buildah, because the version seemed newer than
Buildah's v0.0.0 with a newer commit).

The referenced package is never even physically vendored in here, so remove the
reference:

> go mod edit -droprequire=github.com/openshift/api
> make vendor

Signed-off-by: Miloslav Trmač <mitr@redhat.com>